Add io-package-style Reader and Writer interfaces.
